+++ Seetube/Seetube/Presentation/Youtuber/Common/ViewController/Protocol/SceneResultViewModelBindable.swift
@@ -61,7 +61,7 @@ extension SceneResultViewModelBindable {
             .disposed(by: self.disposeBag)
     }
     
-    func bindPlayingInterval(_ interval: Driver<(start: Int, end: Int)>) {
+    func bindPlayingInterval(_ interval: Driver<(start: Float, end: Float)>) {
         interval
             .drive(with: self) { obj, interval in
                 obj.playInterval(start: interval.start,
